+++ app/src/main/java/com/example/diplomka/DataPoint.java
@@ -3,12 +3,12 @@ package com.example.diplomka;
 public class DataPoint {
 
     public int id;
-    public int dt;
+    public Long dt;
     public float lat;
     public float lon;
     public float noise;
 
-    public DataPoint(int id, int dt, float lat, float lon, float noise) {
+    public DataPoint(int id, long dt, float lat, float lon, float noise) {
         this.id = id;
         this.dt = dt;
         this.lat = lat;
@@ -16,7 +16,7 @@ public class DataPoint {
         this.noise = noise;
     }
 
-    public DataPoint(int dt, float lat, float lon, float noise) {
+    public DataPoint(long dt, float lat, float lon, float noise) {
         this.id = 0;
         this.dt = dt;
         this.lat = lat;
